move run_exe() to separate file
